fix(profiling): add explicit casts to avoid narrowing conversions
Use explicit static_cast where implicit narrowing conversions occur:
- danger.cc: Cast page calculations to safe_memcpy_return_t
- mirrors.cc: Cast table_size calculation to ssize_t
- greenlets.cc: Cast count to int before adding
- stack_renderer.cpp: Cast microsecond_t to int64_t for push_* calls
- vm.cc: Cast size_t to off_t for ftruncate, riovcnt to int for pwritev
- uploader.cpp: Cast len to std::streamsize for write()
- frame.cc: Cast offsetof result to int
- sampler.cpp: Cast uint64_t to double before subtraction

Fixes clang-tidy bugprone-narrowing-conversions warnings.

+++ ddtrace/internal/datadog/profiling/stack/src/echion/mirrors.cc
@@ -10,7 +10,7 @@ MirrorSet::create(PyObject* set_addr)
     }
 
     auto size = set.mask + 1;
-    ssize_t table_size = size * sizeof(setentry);
+    ssize_t table_size = static_cast<ssize_t>(size * sizeof(setentry));
     if (table_size < 0 || table_size > MAX_MIRROR_SIZE) {
         return ErrorKind::MirrorError;
     }
